requirements.txt
python-telegram-bot==20.0b0
Flask
beautifulsoup4
requests
googlesearch-python
gunicorn
hypercorn